Update Python rate.rate field type
The `rate` column of the `rates` table is a NUMERIC in the DB, but was
defined as `db.Float` in Python, which meant alembic kept wanting to
change it in a migration.

Unlike migrating text types, this feels slightly more risky in that it
could affect rounding on numbers. So let's just update the Python
representation. We have other numeric fields, so it's not like we are
particularly consistent.

+++ app/models.py
@@ -1875,7 +1875,7 @@ class Rate(db.Model):
 
     id = db.Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     valid_from = db.Column(db.DateTime, nullable=False)
-    rate = db.Column(db.Float(asdecimal=False), nullable=False)
+    rate = db.Column(db.Numeric(asdecimal=False), nullable=False)
     notification_type = db.Column(notification_types, index=True, nullable=False)
 
     def __str__(self):
